+++ backend/src/routes/oauthRouter.js
@@ -1,6 +1,6 @@
 import { Router } from 'express';
 import OAuthController from '../controllers/oauthController.js';
-import oauth2Middleware from '../lib/oauth2Middleware.js';
+import oauth2Middleware from '../lib/oauth2/oauth2Middleware.js';
 
 const oauthRouter = Router();
